Separate out tests.
